+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [#13] - Add selection to move the cursor to right in the effect columns (thanks [@jimbo1qaz])
 - [#13] - Add highlighted step jump shortcuts (thanks [@jimbo1qaz])
 - Add general settings descriptions in configuration dialog
+- [#13] - Add editable step (thanks [@jimbo1qaz])
 
 ### Fixed
 - [#28] - Fix error when compiling before Qt5.10 (thanks [@papiezak] and others)